// Angular/client-app/src/app/features/cart/models/order.ts
import { OrderItem } from "./orderItem";
import { OrderStatus } from "../../admin-panel/models/orderStatus";

export interface Order{
    id: string;
    name: string;
    description: string | null;
    phone: string;
    email: string | null;
    createdAt: Date;
    items: OrderItem[];
    totalPrice: number;
    totalDiscount: number;
    disabled: boolean;
    priceWithDiscount: number;
    statusId: string;
    status: OrderStatus;
}